+++ client/src/modules/AutoTest/components/index.tsx
@@ -1,5 +1,3 @@
-export { default as AutoTests, type AutoTestsProps } from './AutoTests/AutoTests';
-export { default as TestCard } from './TestCard/TestCard';
-export { default as TestCardColumn } from './TestCardColumn/TestCardColumn';
-export { default as TestDeadlineDate } from './TestDeadlineDate/TestDeadlineDate';
-export { default as Task } from './Task/Task';
+export { default as TaskCard } from './TaskCard/TaskCard';
+export { default as TaskCardColumn } from './TaskCardColumn/TaskCardColumn';
+export { default as TaskDeadlineDate } from './TaskDeadlineDate/TaskDeadlineDate';
