+++ client/src/modules/AutoTest/pages/index.tsx
@@ -1,0 +1,2 @@
+export { default as AutoTests, type AutoTestsProps } from './AutoTests/AutoTests';
+export { default as Task } from './Task/Task';
