+++ client/src/pages/course/student/new-auto-test/index.tsx
@@ -1,5 +1,4 @@
 import { SessionProvider } from 'modules/Course/contexts';
-import { AutoTests, AutoTestsProps } from 'modules/AutoTest/components';
 import { CourseRole } from 'components/withSession';
 import { GetServerSideProps } from 'next';
 import { CoursesTasksApi, CourseTaskDtoTypeEnum, CreateCourseTaskDtoCheckerEnum, ProfileCourseDto } from 'api';
@@ -7,6 +6,7 @@ import { getTokenFromContext } from 'utils/server';
 import { noAccessResponse, notAuthorizedResponse } from 'modules/Course/data';
 import { UserService } from 'services/user';
 import { getApiConfiguration } from 'utils/axios';
+import { AutoTests, AutoTestsProps } from 'modules/AutoTest/pages';
 
 export const getServerSideProps: GetServerSideProps<{ course: ProfileCourseDto }> = async ctx => {
   try {
